added launches and events to store and update article endpoint

+++ app/Http/Controllers/ArticleController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Article;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -12,7 +13,7 @@ class ArticleController extends Controller
 
     public function index(Request $request) {
         $skip_entries = $request->page * self::PAGE_SIZE;
-        return Article::orderBy("id")
+        return Article::orderBy('id')
             ->skip($skip_entries)
             ->take(self::PAGE_SIZE)
             ->with('launches')
@@ -44,18 +45,17 @@ class ArticleController extends Controller
             'events'
         ]);
 
-        return Article::createFromExternalArticle($external_article);
+        $new_article = Article::createFromExternalArticle($external_article);
+
+        return $this->getArticleWithLaunchesAndEvents($new_article);
     }
 
     public function show(Article $article) {
-        return Article::where('id', $article->id)
-            ->with('launches')
-            ->with('events')
-            ->first();
+        return $this->getArticleWithLaunchesAndEvents($article);
     }
 
     public function update(Request $request, Article $article) {
-        $article->update($request->only([
+        $fields_to_update = $request->only([
             'title',
             'url',
             'imageUrl',
@@ -64,19 +64,36 @@ class ArticleController extends Controller
             'publishedAt',
             'updatedAt',
             'featured'
-        ]));
+        ]);
 
-        if(isset($request->launches)){
+        if($fields_to_update['publishedAt']){
+            $fields_to_update['publishedAt'] = Carbon::parse($fields_to_update['publishedAt']);
+        }
+
+        if($fields_to_update['updatedAt']){
+            $fields_to_update['updatedAt'] = Carbon::parse($fields_to_update['updatedAt']);
+        }
+
+        $article->update($fields_to_update);
+
+        if(isset($request->launches) && count($request->launches) > 0){
             DB::table('articles_launches')->where('article_id', $article->id)->delete();
         }
 
-        if(isset($request->events)){
+        if(isset($request->events) && count($request->events) > 0){
             DB::table('articles_events')->where('article_id', $article->id)->delete();
         }
 
         $article->associateLaunchesAndEvents($request->launches, $request->events);
 
-        return $article;
+        return $this->getArticleWithLaunchesAndEvents($article);
+    }
+
+    private function getArticleWithLaunchesAndEvents($article){
+        return Article::where('id', $article->id)
+            ->with('launches')
+            ->with('events')
+            ->first();
     }
 
     public function destroy(Article $article) {
